--- docs/src/components/MdxContent.tsx
@@ -1,31 +1,0 @@
-/* (c) Copyright Frontify Ltd., all rights reserved. */
-
-import * as fondueComponents from '@frontify/fondue-components';
-import { getMDXComponent } from 'mdx-bundler/client';
-import { useMemo } from 'react';
-
-import { EditableComponent } from './EditableComponent/EditableComponent';
-
-export const MdxContent = ({ data }: { data: string }) => {
-    const Component = useMemo(() => getMDXComponent(data), [data]);
-
-    return (
-        <Component
-            components={{
-                ...fondueComponents,
-                code: (props) => {
-                    if (props.children) {
-                        return (
-                            <EditableComponent
-                                language={props.className?.split('language-')[1]}
-                                code={props.children as string}
-                            />
-                        );
-                    } else {
-                        throw new Error('Invalid Code Content');
-                    }
-                },
-            }}
-        />
-    );
-};
